+++ src/main/java/de/maxhenkel/enhancedgroups/EnhancedGroupsVoicechatPlugin.java
@@ -1,11 +1,13 @@
 package de.maxhenkel.enhancedgroups;
 
 import de.maxhenkel.enhancedgroups.config.PersistentGroup;
+import de.maxhenkel.enhancedgroups.events.AutoJoinGroupEvents;
 import de.maxhenkel.voicechat.api.Group;
 import de.maxhenkel.voicechat.api.VoicechatApi;
 import de.maxhenkel.voicechat.api.VoicechatPlugin;
 import de.maxhenkel.voicechat.api.VoicechatServerApi;
 import de.maxhenkel.voicechat.api.events.EventRegistration;
+import de.maxhenkel.voicechat.api.events.PlayerConnectedEvent;
 import de.maxhenkel.voicechat.api.events.VoicechatServerStartedEvent;
 
 import javax.annotation.Nullable;
@@ -29,6 +31,7 @@ public class EnhancedGroupsVoicechatPlugin implements VoicechatPlugin {
     @Override
     public void registerEvents(EventRegistration registration) {
         registration.registerEvent(VoicechatServerStartedEvent.class, this::onServerStarted);
+        registration.registerEvent(PlayerConnectedEvent.class, AutoJoinGroupEvents::onPlayerConnected);
     }
 
     private void onServerStarted(VoicechatServerStartedEvent event) {