+++ src/main/java/de/maxhenkel/enhancedgroups/command/PersistentGroupCommands.java
@@ -20,7 +20,6 @@ import net.minecraft.network.chat.ComponentUtils;
 import net.minecraft.network.chat.HoverEvent;
 
 import javax.annotation.Nullable;
-import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 
@@ -64,9 +63,21 @@ public class PersistentGroupCommands {
         literalBuilder.then(Commands.literal("add").then(nameArg));
 
         literalBuilder.then(Commands.literal("remove").then(Commands.argument("id", UuidArgument.uuid()).executes(context -> {
-            return removePersistentGroup(context, UuidArgument.getUuid(context, "id"));
+            UUID id = UuidArgument.getUuid(context, "id");
+            PersistentGroup group = EnhancedGroups.PERSISTENT_GROUP_STORE.getGroup(id);
+            if (group == null) {
+                context.getSource().sendFailure(Component.literal("Group not found or not persistent"));
+                return 0;
+            }
+            return removePersistentGroup(context, group);
         })).then(Commands.argument("name", StringArgumentType.string()).executes(context -> {
-            return removePersistentGroup(context, StringArgumentType.getString(context, "name"));
+            String name = StringArgumentType.getString(context, "name");
+            PersistentGroup group = EnhancedGroups.PERSISTENT_GROUP_STORE.getGroup(name);
+            if (group == null) {
+                context.getSource().sendFailure(Component.literal("Group not found or not persistent"));
+                return 0;
+            }
+            return removePersistentGroup(context, group);
         })));
 
         literalBuilder.then(Commands.literal("list").executes(PersistentGroupCommands::listPersistentGroups));
@@ -96,68 +107,35 @@ public class PersistentGroupCommands {
         return 1;
     }
 
-    public static int removePersistentGroup(CommandContext<CommandSourceStack> commandSource, String name) throws CommandSyntaxException {
+    public static int removePersistentGroup(CommandContext<CommandSourceStack> commandSource, PersistentGroup persistentGroup) throws CommandSyntaxException {
         if (EnhancedGroupsVoicechatPlugin.SERVER_API == null) {
             commandSource.getSource().sendFailure(Component.literal("Voice chat not connected"));
-            return 1;
+            return 0;
         }
 
-        Collection<Group> groups = EnhancedGroupsVoicechatPlugin.SERVER_API.getGroups();
+        UUID voicechatId = EnhancedGroups.PERSISTENT_GROUP_STORE.getVoicechatId(persistentGroup.getId());
 
-        int removedCount = 0;
-
-        for (Group group : groups) {
-            if (group.getName().equals(name)) {
-                boolean removed = EnhancedGroupsVoicechatPlugin.SERVER_API.removeGroup(group.getId());
-                if (removed) {
-                    PersistentGroup cachedGroup = EnhancedGroups.PERSISTENT_GROUP_STORE.getCached(group.getId());
-                    if (cachedGroup == null) {
-                        commandSource.getSource().sendFailure(Component.literal("This group was not created by EnhancedGroups"));
-                        return 1;
-                    }
-                    EnhancedGroups.PERSISTENT_GROUP_STORE.removeGroup(cachedGroup);
-                    commandSource.getSource().sendSuccess(Component.literal("Removed group " + name), false);
-                } else {
-                    commandSource.getSource().sendFailure(Component.literal("Could not remove group " + name));
-                }
-                removedCount++;
-            }
+        if (voicechatId == null) {
+            commandSource.getSource().sendFailure(Component.literal("Group not found or not persistent"));
+            return 0;
         }
 
-        if (removedCount <= 0) {
-            commandSource.getSource().sendFailure(Component.literal("Could not find group " + name));
-        }
-
-        return 1;
-    }
-
-    public static int removePersistentGroup(CommandContext<CommandSourceStack> commandSource, UUID id) throws CommandSyntaxException {
-        if (EnhancedGroupsVoicechatPlugin.SERVER_API == null) {
-            commandSource.getSource().sendFailure(Component.literal("Voice chat not connected"));
-            return 1;
-        }
-
-        Group group = EnhancedGroupsVoicechatPlugin.SERVER_API.getGroup(id);
+        Group group = EnhancedGroupsVoicechatPlugin.SERVER_API.getGroup(voicechatId);
 
         if (group == null) {
             commandSource.getSource().sendFailure(Component.literal("Group not found"));
-            return 1;
+            return 0;
         }
 
-        boolean removed = EnhancedGroupsVoicechatPlugin.SERVER_API.removeGroup(id);
+        boolean removed = EnhancedGroupsVoicechatPlugin.SERVER_API.removeGroup(voicechatId);
         if (removed) {
-            PersistentGroup cachedGroup = EnhancedGroups.PERSISTENT_GROUP_STORE.getCached(group.getId());
-            if (cachedGroup == null) {
-                commandSource.getSource().sendFailure(Component.literal("This group was not created by EnhancedGroups"));
-                return 1;
-            }
-            EnhancedGroups.PERSISTENT_GROUP_STORE.removeGroup(cachedGroup);
-            commandSource.getSource().sendSuccess(Component.literal("Removed group " + group.getName()), false);
+            EnhancedGroups.PERSISTENT_GROUP_STORE.removeGroup(persistentGroup);
+            commandSource.getSource().sendSuccess(Component.literal("Removed group %s".formatted(group.getName())), false);
+            return 1;
         } else {
-            commandSource.getSource().sendFailure(Component.literal("Could not remove group " + group.getName()));
+            commandSource.getSource().sendFailure(Component.literal("Could not remove group %s".formatted(group.getName())));
+            return 0;
         }
-
-        return 1;
     }
 
     public static int listPersistentGroups(CommandContext<CommandSourceStack> commandSource) throws CommandSyntaxException {
@@ -166,23 +144,33 @@ public class PersistentGroupCommands {
             return 1;
         }
 
-        List<Group> groups = EnhancedGroupsVoicechatPlugin.SERVER_API.getGroups().stream().filter(Group::isPersistent).toList();
+        List<PersistentGroup> groups = EnhancedGroups.PERSISTENT_GROUP_STORE.getGroups();
 
         if (groups.isEmpty()) {
             commandSource.getSource().sendSuccess(Component.literal("There are no persistent groups"), false);
         }
 
-
-        for (Group group : groups) {
-            PersistentGroup cachedGroup = EnhancedGroups.PERSISTENT_GROUP_STORE.getCached(group.getId());
-            if (cachedGroup == null) {
-                continue;
-            }
-            commandSource.getSource().sendSuccess(Component.literal(group.getName()).append(" ").append(ComponentUtils.wrapInSquareBrackets(Component.literal("Remove")).withStyle(style -> {
-                return style.withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/" + PERSISTENTGROUP_COMMAND + " remove " + group.getId()))
-                        .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal("Click to remove group")))
-                        .applyFormat(ChatFormatting.GREEN);
-            })), false);
+        for (PersistentGroup group : EnhancedGroups.PERSISTENT_GROUP_STORE.getGroups()) {
+            commandSource.getSource().sendSuccess(Component.literal(group.getName())
+                            .append(" ")
+                            .append(ComponentUtils.wrapInSquareBrackets(Component.literal("Remove")).withStyle(style -> {
+                                return style.withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/" + PERSISTENTGROUP_COMMAND + " remove " + group.getId()))
+                                        .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal("Click to remove group")))
+                                        .applyFormat(ChatFormatting.GREEN);
+                            }))
+                            .append(" ")
+                            .append(ComponentUtils.wrapInSquareBrackets(Component.literal("Auto Join")).withStyle(style -> {
+                                return style.withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/" + AutoJoinGroupCommands.AUTOJOINGROUP_COMMAND + " set " + group.getId()))
+                                        .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal("Click to automatically connect to the group when joining")))
+                                        .applyFormat(ChatFormatting.GREEN);
+                            }))
+                            .append(" ")
+                            .append(ComponentUtils.wrapInSquareBrackets(Component.literal("Copy ID")).withStyle(style -> {
+                                return style.withClickEvent(new ClickEvent(ClickEvent.Action.COPY_TO_CLIPBOARD, group.getId().toString()))
+                                        .withHoverEvent(new HoverEvent(HoverEvent.Action.SHOW_TEXT, Component.literal("Click to copy group ID")))
+                                        .applyFormat(ChatFormatting.GREEN);
+                            }))
+                    , false);
         }
         return groups.size();
     }